* refs/pull/20527/head:
	mds: various fixes for backport
	mds: convert to boost::string_view
	test/encoding: refactor to avoid escaping shell magic
	mds: minor refactor of SimpleLock
	mds: track Capability in mempool
	mds: move CInode container members to mempool
	mds: move CDentry container members to mempool
	mds: move CDir container members to mempool
	mds: put MDSCacheObject compact_map in mempool
	common: use size_t for object size
	mds: convert to allocator agnostic string_view
	mds: simplify initialization
	compact_*: support mempool allocated containers

+++ src/common/escape.h
@@ -15,6 +15,8 @@
 #ifndef CEPH_RGW_ESCAPE_H
 #define CEPH_RGW_ESCAPE_H
 
+#include <stdlib.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -22,7 +24,7 @@ extern "C" {
 /* Returns the length of a buffer that would be needed to escape 'buf'
  * as an XML attrribute
  */
-int escape_xml_attr_len(const char *buf);
+size_t escape_xml_attr_len(const char *buf);
 
 /* Escapes 'buf' as an XML attribute. Assumes that 'out' is at least long
  * enough to fit the output. You can find out the required length by calling
@@ -33,13 +35,13 @@ void escape_xml_attr(const char *buf, char *out);
 /* Returns the length of a buffer that would be needed to escape 'buf'
  * as an JSON attrribute
  */
-int escape_json_attr_len(const char *buf, int src_len);
+size_t escape_json_attr_len(const char *buf, size_t src_len);
 
 /* Escapes 'buf' as an JSON attribute. Assumes that 'out' is at least long
  * enough to fit the output. You can find out the required length by calling
  * escape_json_attr_len first.
  */
-void escape_json_attr(const char *buf, int src_len, char *out);
+void escape_json_attr(const char *buf, size_t src_len, char *out);
 
 /* Note: we escape control characters. Although the XML spec doesn't actually
  * require this, Amazon does it in their XML responses.
